Renamed and componentized Share access form sections
