+++ ui/src/AccountDropdown/GrantEditAccessConfirmationForm.scss
