+++ ui/src/AccountDropdown/InviteEditorsFormSection.scss
@@ -1,0 +1,69 @@
+/*!
+ * Copyright (c) 2020 Ford Motor Company
+ * All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@import "../Application/Styleguide/Colors.scss";
+
+.inviteEditorsForm {
+  width: 415px;
+  margin: 0 auto;
+
+  .editTitle {
+    color: $purple-title;
+  }
+
+  .inviteEditorsLabel {
+    text-align: left;
+    font-style: normal;
+    font-weight: bold;
+    font-size: 14px;
+    line-height: 16px;
+    color: $gray-1;
+  }
+
+  .emailTextarea {
+    font-family: inherit;
+    font-weight: inherit;
+    font-size: 14px;
+    line-height: 16px;
+    height: 36px;
+    width: 100%;
+    box-sizing: border-box;
+    resize: none;
+    padding: 8px;
+    margin: 12px 0 32px;
+    background: #F2F2F2;
+    color: $gray-1;
+    border: 1px solid transparent;
+    border-radius: 6px;
+
+    &::placeholder {
+      color: $gray-3;
+    }
+  }
+
+  .buttonsContainer {
+    text-align: center;
+    height: 32px;
+
+    .cancelButton {
+        margin-right: 16px;
+    }
+  }
+}
+
+
+
