+++ ui/src/AccountDropdown/ShareAccessButton.tsx
@@ -26,7 +26,7 @@ interface Props {
 }
 
 function ShareAccessButton({ setCurrentModal }: Props): JSX.Element {
-    const openEditContributorsModal = (): void => setCurrentModal({modal: AvailableModals.EDIT_CONTRIBUTORS});
+    const openEditContributorsModal = (): void => setCurrentModal({modal: AvailableModals.SHARE_SPACE_ACCESS});
     const showButton = window.runConfig.invite_users_to_space_enabled;
     const onkeydown = (event: React.KeyboardEvent): void => {
         if (event.key === 'Enter') openEditContributorsModal();
