+++ ui/src/AccountDropdown/ViewOnlyAccessFormSection.scss
@@ -1,4 +1,4 @@
-/*!
+/*
  * Copyright (c) 2020 Ford Motor Company
  * All rights reserved.
  *
@@ -17,54 +17,11 @@
 
 @import "../Application/Styleguide/Colors.scss";
 
-.editContributorsContainer {
+.viewOnlyAccessForm {
   width: 415px;
   margin: 0 auto;
 
-  label {
-    text-align: left;
-    font-style: normal;
-    font-weight: bold;
-    font-size: 14px;
-    line-height: 16px;
-    color: $gray-1;
-  }
-
-  textarea {
-    font-family: inherit;
-    font-weight: inherit;
-    font-size: 12px;
-    color: $gray-1;
-
-    border: 1px solid $gray-4;
-    box-sizing: border-box;
-    border-radius: 2px;
-
-    height: 64px;
-    width: 100%;
-    resize: none;
-    padding: 8px;
-    margin-top: 12px;
-    margin-bottom: 32px;
-  }
-
-  .editContributorsButtonContainer {
-    width: 100%;
-    height: 32px;
-
-    display: flex;
-    justify-content: space-between;
-
-    button {
-      flex: 1;
-
-      &.editContributorsCancelButton {
-        margin-right: 16px;
-      }
-    }
-  }
-
-  .readOnlySwitchContainer {
+  .viewOnlyToggleContainer {
     display: inline-block;
     margin-bottom: 24px;
 
@@ -180,10 +137,6 @@
       }
     }
   }
-
-  .editTitle {
-    color: $purple-title;
-  }
 }
 
 .react-switch-bg {
@@ -195,6 +148,3 @@
     background: #E1B7E5 !important;
   }
 }
-
-
-
